double -> float

+++ src/main/java/com/linngdu664/bsf3lite/entity/golem/AbstractBSFSnowGolemEntity.java
@@ -204,8 +204,8 @@ public abstract class AbstractBSFSnowGolemEntity extends PathfinderMob implement
         return getBoundingBox().inflate(searchDistance, searchDistance, searchDistance);
     }
 
-    public Vec3 getMiddleModelForward(float partialTicks, double degreeOffset) {
-        return BSFCommonUtil.radRotationToVector(1, (Mth.lerp(partialTicks, this.yBodyRotO + ((this.yHeadRotO - this.yBodyRotO) * 0.25), this.yBodyRot + ((this.yHeadRot - this.yBodyRot) * 0.25)) + 90 + degreeOffset) * Mth.DEG_TO_RAD, 0);
+    public Vec3 getMiddleModelForward(float partialTicks, float degreeOffset) {
+        return BSFCommonUtil.radRotationToVector(1, (Mth.lerp(partialTicks, this.yBodyRotO + ((this.yHeadRotO - this.yBodyRotO) * 0.25f), this.yBodyRot + ((this.yHeadRot - this.yBodyRot) * 0.25f)) + 90f + degreeOffset) * Mth.DEG_TO_RAD, 0);
     }
 
     @Override
